# Conflicts:
#	zkingedu/src/main/java/com/zking/zkingedu/common/controller/QQLogin.java
#	zkingedu/src/main/java/com/zking/zkingedu/common/dao/UserDao.java
#	zkingedu/src/main/resources/rebel.xml

+++ zkingedu/src/main/java/com/zking/zkingedu/common/model/Charge.java
@@ -19,7 +19,7 @@ public class Charge implements Serializable {
     //用户id
     private Integer chargeUid;
     //金额
-    private Double chargeMoney;
+    private Integer chargeMoney;
     //积分
     private Integer chargeIntegral;
     //'充值时间
